move classes around, write meta inside the download function, and fix existence checking
