+++ vodbot/twitch.py
@@ -1,10 +1,77 @@
 # Module to make API calls to Twitch.tv
 
-from .channel import Channel
-from .clip import Clip
-from .video import Video
-
 import requests
+import json
+
+class Vod:
+	def __init__(self, json):
+		self.id = json["id"] # url: twitch.tv/videos/{id}
+		self.user_id = json["user_id"]
+		self.user_name = json["user_name"]
+		self.title = json["title"]
+		self.created_at = json["created_at"]
+		self.duration = json["duration"]
+		
+		self.url = f"twitch.tv/videos/{self.id}"
+	
+	def __repr__(self):
+		return f"VOD({self.id}, {self.created_at}, {self.user_name}, {self.created_at}, {self.duration})"
+	
+	def write_meta(self, filename):
+		jsondict = {
+			"id": self.id,
+			"user_id": self.user_id,
+			"user_name": self.user_name,
+			"title": self.title,
+			"created_at": self.created_at,
+			"duration": self.duration
+		}
+		
+		with open(filename, "w") as f:
+			json.dump(jsondict, f, sort_keys=True, indent=4)
+
+
+class Clip:
+	def __init__(self, json):
+		self.id = json["id"] # url: twitch.tv/videos/{id}
+		self.user_id = json["broadcaster_id"]
+		self.user_name = json["broadcaster_name"]
+		self.clipper_id = json["creator_id"]
+		self.clipper_name = json["creator_name"]
+		self.title = json["title"]
+		self.created_at = json["created_at"]
+		self.view_count = json["view_count"]
+		
+		self.url = f"twitch.tv/{self.user_name}/clip/{self.id}"
+	
+	def __repr__(self):
+		return f"Clip({self.title}, {self.created_at}, {self.user_name}, {self.clipper_name})"
+	
+	def write_meta(self, filename):
+		jsondict = {
+			"id": self.id,
+			"user_id": self.user_id,
+			"user_name": self.user_name,
+			"clipper_id": self.clipper_id,
+			"clipper_name": self.clipper_name,
+			"title": self.title,
+			"created_at": self.created_at,
+			"view_count": self.view_count
+		}
+		
+		with open(filename, "w") as f:
+			json.dump(jsondict, f, sort_keys=True, indent=4)
+
+
+class Channel:
+	def __init__(self, json):
+		self.id = json["id"]
+		self.login = json["login"]
+		self.display_name = json["display_name"]
+		self.created_at = json["created_at"]
+	
+	def __repr__(self):
+		return f"Channel({self.id}, {self.display_name})"
 
 
 def get_access_token(CLIENT_ID, CLIENT_SECRET):
@@ -124,7 +191,7 @@ def _get_channel_content(video_url, noun, channel, headers):
 			# Live VODs don't have thumbnails
 			if vod["thumbnail_url"] != "":
 				if noun == "VOD":
-					videos.append(Video(vod))
+					videos.append(Vod(vod))
 				elif noun == "Clip":
 					videos.append(Clip(vod))
